backend/requirements.txt
# Core Dependencies
fastapi==0.95.2
pydantic==1.10.13
uvicorn==0.15.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.1
pyyaml==6.0.2
aiohttp==3.11.18
requests==2.31.0

# AI and ML
openai==1.3.0
tenacity==8.2.3
shap==0.44.0
scikit-learn==1.4.2
matplotlib==3.8.4
numpy==1.26.4
ollama-python==0.1.6

# Healthcare
fhirclient==4.3.1

# Monitoring and Logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7

# Security
cryptography==42.0.2

# Utilities
aiofiles==23.2.1
click==8.1.7
rich==13.7.0
tabulate==0.9.0
redis==5.0.1
cachetools==5.3.2
